+++ tgui/packages/common/collections.ts
@@ -174,6 +174,13 @@ export const sortBy = (...iterateeFns) => array => {
 export const sort = sortBy();
 
 /**
+ * Returns a range of numbers from start to end, exclusively.
+ * For example, range(0, 5) will return [0, 1, 2, 3, 4].
+ */
+export const range = (start: number, end: number): number[] =>
+  new Array(end - start).fill(null).map((_, index) => index + start);
+
+/**
  * A fast implementation of reduce.
  */
 export const reduce = (reducerFn, initialValue) => array => {
