+++ tgui/packages/common/math.ts
@@ -98,10 +98,3 @@ export const numberOfDecimalDigits = value => {
   }
   return 0;
 };
-
-/**
- * Returns a range of numbers from start to end, exclusively.
- * For example, range(0, 5) will return [0, 1, 2, 3, 4].
- */
-export const range = (start: number, end: number): number[] =>
-  new Array(end - start).fill(null).map((_, index) => index + start);
