+++ tgui/packages/tgui/interfaces/StripMenu.tsx
@@ -1,4 +1,4 @@
-import { range } from "common/math";
+import { range } from "common/collections";
 import { BooleanLike } from "common/react";
 import { resolveAsset } from "../assets";
 import { useBackend } from "../backend";
